Move gallery logic to util
